requirements.txt
# Core Flask dependencies
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23

# PDF processing
PyPDF2==2.10.5
reportlab==3.6.1

# Excel export
xlsxwriter==3.0.3

# Google Drive API
google-api-python-client==2.25.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6

# Environment management
python-dotenv==0.19.0

# Production server
gunicorn==20.1.0

# These need to be installed separately with special options on some systems
# PyMuPDF is imported as 'fitz'
# Use these commands if pip install fails:
# pip install --upgrade pip
# pip install --upgrade wheel
# pip install --upgrade setuptools
# pip install PyMuPDF==1.18.0